+++ TidalPy/__init__.py
@@ -50,10 +50,5 @@ else:
     if not os.path.exists(world_config_loc):
         os.makedirs(world_config_loc)
 
-# Bring up various functionality to top-level
-#    Performance / Basic functionality
-from .utilities.performance import clear_cache
-#    Physics
-
-#    OOP
-from .structures import build_world, build_from_world, scale_from_world, Orbit
+# Import package functions
+from .cache import clear_cache